requirements.txt
streamlit
groq
langchain
groqcloud
python-dotenv
PyMuPDF
pydantic
pdfminer.six
spacy
scikit-learn
reportlab
